third_party/blink/web_tests/virtual/mutation-events-enabled/dom/mutation-event-tests/fast/dom/Range/range-created-during-remove-children-expected.txt

CONSOLE WARNING: Listener added for a 'DOMNodeRemovedFromDocument' mutation event. This event type is no longer supported, and will be removed from this browser VERY soon. Consider using MutationObserver instead. See https://chromestatus.com/feature/5083947249172480 for more information.
PASS ranges["blur"].startOffset is 0
PASS ranges["blur"].endOffset is 0
PASS ranges["DOMNodeRemovedFromDocument"].startOffset is 0
PASS ranges["DOMNodeRemovedFromDocument"].endOffset is 0
PASS successfullyParsed is true

TEST COMPLETE
